+++ app/db_gui/view/db_gui_application.rb
@@ -13,7 +13,7 @@ class DbGui
           
       before_body do
         @db = Model::Db.new
-#         menu_bar # TODO implement
+        menu_bar
       end
   
       body {
@@ -35,38 +35,48 @@ class DbGui
         }
       }
   
-#       def menu_bar
-#         menu('File') {
-#           menu_item('Preferences...') {
-#             on_clicked do
-#               display_preferences_dialog
-#             end
-#           }
-#
-#           quit_menu_item if OS.mac?
-#         }
-#         menu('Help') {
-#           if OS.mac?
-#             about_menu_item {
-#               on_clicked do
-#                 display_about_dialog
-#               end
-#             }
-#           end
-#
-#           menu_item('About') {
-#             on_clicked do
-#               display_about_dialog
-#             end
-#           }
-#         }
-#       end
-#
-#       def display_about_dialog
-#         message = "Db Gui #{VERSION}\n\n#{LICENSE}"
-#         msg_box('About', message)
-#       end
-#
+      def menu_bar
+        menu('Edit') {
+          menu_item('Copy Table') {
+            enabled <= [db, :db_command_result_rows, computed_by: :db_command_result, on_read: -> (data) { !data.empty? }]
+            
+            on_clicked do
+              db.copy_table
+            end
+          }
+
+          menu_item('Copy Selected Row') {
+            enabled <= [db, :db_command_result_rows, computed_by: :db_command_result, on_read: -> (data) { !data.empty? }]
+            
+            on_clicked do
+              db.copy_selected_row
+            end
+          }
+
+          quit_menu_item if OS.mac?
+        }
+        menu('Help') {
+          if OS.mac?
+            about_menu_item {
+              on_clicked do
+                display_about_dialog
+              end
+            }
+          end
+
+          menu_item('About') {
+            on_clicked do
+              display_about_dialog
+            end
+          }
+        }
+      end
+
+      def display_about_dialog
+        message = "DB GUI #{VERSION}\n\n#{LICENSE}"
+        msg_box('About', message)
+      end
+
 #       def display_preferences_dialog
 #         window {
 #           title 'Preferences'